Replace TimedRotatingFileHandler with RotatingFileHandler
logger_config previously set a TimedRotatingFileHandler in an attempt to
automatically rotate log files. However, the runtime of most scripts was
less than the rotate interval, so logs were hardly ever rotated. The
RotatingFileHandler works on filesize instead of time since last
rotation, so it works better with short-runtime scripts.

maxBytes is set to 10 MiB, after which the log file will be rotated.

Closes #2