Don't switch logging FileHandler based on environment
Previously, either FileHandler or a RotatingFileHandler was used
depending on if on_toolforge was True. This was to prevent unwanted log
rotation during development, based on an expectation that
TimedRotatingFileHandler would rotate logs after the interval had
elapsed, even if the program wasn't running the whole time. That wasn't
the case, but now that logger_config uses RotatingFileHandler it is
definitely unnecessary.